requirements.txt
pafy==0.5.4
wget==3.2
Pillow==6.1.0
